ToolBar transferred to ToolBarUI and more
