+++ app/src/main/java/dev/msartore/gallery/MainActivity.kt
@@ -6,7 +6,6 @@ import android.database.ContentObserver
 import android.net.Uri
 import android.os.Build
 import android.os.Bundle
-import android.provider.MediaStore
 import android.provider.Settings
 import android.util.Log
 import androidx.activity.ComponentActivity
@@ -15,37 +14,31 @@ import androidx.activity.compose.setContent
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.IntentSenderRequest
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.animation.*
+import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.scaleIn
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.CircularProgressIndicator
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.ArrowBack
-import androidx.compose.material.icons.rounded.Close
-import androidx.compose.material.icons.rounded.Delete
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.*
+import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.mutableStateListOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.lifecycle.lifecycleScope
 import coil.annotation.ExperimentalCoilApi
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import dev.msartore.gallery.ui.compose.FileAndMediaPermission
 import dev.msartore.gallery.ui.compose.ImageViewUI
 import dev.msartore.gallery.ui.compose.MediaListUI
-import dev.msartore.gallery.ui.compose.basic.Icon
+import dev.msartore.gallery.ui.compose.ToolBarUI
 import dev.msartore.gallery.ui.theme.GalleryTheme
 import dev.msartore.gallery.utils.*
 import kotlinx.coroutines.delay
@@ -79,7 +72,10 @@ class MainActivity : ComponentActivity() {
                 updateNeeded = true
             }
 
-            counterImageToDelete -= 1
+            if (counterImageToDelete > 0)
+                counterImageToDelete -= 1
+
+            Log.d("MainActivity", "counterImageToDelete: $counterImageToDelete updateNeeded: $updateNeeded")
 
             if (updateNeeded && counterImageToDelete == 0)
                 cor {
@@ -94,14 +90,8 @@ class MainActivity : ComponentActivity() {
 
         val mediaList = mutableStateListOf<MediaClass>()
         val loading = mutableStateOf(false)
-
-        val intentCamera =
-            if (checkCameraHardware(this))
-                Intent(MediaStore.INTENT_ACTION_STILL_IMAGE_CAMERA)
-            else
-                null
-
-
+        val selectedMedia = mutableStateOf<MediaClass?>(null)
+        val mediaDelete = MutableSharedFlow<DeleteMediaVars>()
         val getContent = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
             if (it.resultCode != RESULT_OK) {
                 finishAffinity()
@@ -114,12 +104,25 @@ class MainActivity : ComponentActivity() {
             }
         }
 
-        val selectedMedia = mutableStateOf<MediaClass?>(null)
+        cor {
+            mediaDelete.collect {
+                deleteInProgress = true
+
+                counterImageToDelete = it.listUri.size
+
+                it.listUri.forEach { uri ->
+                    setVarsAndDeleteImage(
+                        uri,
+                        it.action
+                    )
+                }
+            }
+        }
 
         cor {
-
             mediaListFlow.collect {
 
+                Log.d("MediaListFlow", "$deleteInProgress")
                 if (!deleteInProgress)
                     cor {
 
@@ -128,8 +131,6 @@ class MainActivity : ComponentActivity() {
                         mediaList.clear()
                         mediaList.addAll(contentResolver.queryImageMediaStore())
                         mediaList.addAll(contentResolver.queryVideoMediaStore())
-
-                        mediaList.sortBy { it.uri }
 
                         delay(10)
 
@@ -143,8 +144,6 @@ class MainActivity : ComponentActivity() {
             mediaList.clear()
             mediaList.addAll(contentResolver.queryImageMediaStore())
             mediaList.addAll(contentResolver.queryVideoMediaStore())
-
-            mediaList.sortBy { it.uri }
         }
 
         setContent {
@@ -155,14 +154,14 @@ class MainActivity : ComponentActivity() {
                 ) {
                     val scrollState = rememberLazyListState()
                     val checkBoxVisible = remember { mutableStateOf(false) }
+
                     // Remember a SystemUiController
                     val systemUiController = rememberSystemUiController()
                     val useDarkIcons = androidx.compose.material.MaterialTheme.colors.isLight
-                    val systemBarsColor = MaterialTheme.colorScheme.background
 
-                    LaunchedEffect(key1 = useDarkIcons) {
+                    SideEffect {
                         systemUiController.setSystemBarsColor(
-                            color = systemBarsColor,
+                            color = Color.Transparent,
                             darkIcons = useDarkIcons
                         )
                     }
@@ -208,7 +207,6 @@ class MainActivity : ComponentActivity() {
                                     }
 
                                     if (selectedMedia.value == null) {
-
                                         if (loading.value) {
                                             CircularProgressIndicator(
                                                 modifier = Modifier
@@ -239,138 +237,25 @@ class MainActivity : ComponentActivity() {
                                             ImageViewUI(selectedMedia.value!!)
                                         }
                                     }
-                                }
-                            )
 
-                            androidx.compose.animation.AnimatedVisibility(
-                                visible = scrollState.firstVisibleItemScrollOffset == 0 || !scrollState.isScrollInProgress  || checkBoxVisible.value,
-                                enter = slideInVertically(),
-                                exit = slideOutVertically()
-                            ) {
-                                Row(
-                                    modifier = Modifier
-                                        .wrapContentHeight()
-                                        .fillMaxWidth()
-                                        .background(
+                                    ToolBarUI(
+                                        visible = scrollState.firstVisibleItemScrollOffset == 0 || !scrollState.isScrollInProgress  || checkBoxVisible.value,
+                                        mediaList = mediaList,
+                                        mediaDelete = mediaDelete,
+                                        selectedMedia = selectedMedia,
+                                        checkBoxVisible = checkBoxVisible,
+                                        backgroundColor =
                                             if (selectedMedia.value == null) {
                                                 if (scrollState.firstVisibleItemScrollOffset == 0) {
                                                     MaterialTheme.colorScheme.background
                                                 } else {
                                                     MaterialTheme.colorScheme.surface
                                                 }
-                                            } else Color.Transparent,
-                                            RoundedCornerShape(
-                                                bottomEnd = 16.dp,
-                                                bottomStart = 16.dp
-                                            )
-                                        )
-                                        .padding(20.dp),
-                                    verticalAlignment = Alignment.CenterVertically,
-                                    horizontalArrangement = Arrangement.SpaceBetween
-                                ) {
-                                    if (selectedMedia.value != null) {
-                                        Icon(
-                                            imageVector = Icons.Rounded.ArrowBack,
-                                            tint = Color.White,
-                                        ) {
-                                            selectedMedia.value = null
-                                        }
-
-                                        Icon(
-                                            painter = painterResource(id = R.drawable.baseline_share_24),
-                                            tint = Color.White
-                                        ) {
-                                            selectedMedia.value?.uri?.let {
-                                                shareImage(arrayListOf(it))
                                             }
-                                        }
-
-                                        Icon(
-                                            imageVector = Icons.Rounded.Delete,
-                                            tint = Color.White
-                                        ) {
-                                            cor {
-                                                setVarsAndDeleteImage(selectedMedia.value!!.uri) {
-                                                    selectedMedia.value = null
-                                                }
-                                            }
-                                        }
-                                    }
-
-                                    when {
-                                        !checkBoxVisible.value && selectedMedia.value == null -> {
-                                            Text(
-                                                text = "Gallery",
-                                                fontSize = 20.sp,
-                                                textAlign = TextAlign.Center,
-                                                color = Color.DarkGray,
-                                                fontWeight = FontWeight.SemiBold,
-                                                fontFamily = FontFamily.SansSerif
-                                            )
-
-                                            Icon(
-                                                painter = painterResource(id = R.drawable.baseline_photo_camera_24),
-                                                tint = Color.DarkGray,
-                                                shadowEnabled = false
-                                            ) {
-                                                if (intentCamera != null) {
-                                                    startActivitySafely(intentCamera)
-                                                }
-                                            }
-                                        }
-                                        checkBoxVisible.value -> {
-                                            BackHandler(enabled = true){
-                                                mediaList.forEach {
-                                                    it.selected.value = false
-                                                }
-                                                checkBoxVisible.value = false
-                                            }
-
-                                            Icon(
-                                                imageVector = Icons.Rounded.Close
-                                            ) {
-                                                mediaList.forEach {
-                                                    it.selected.value = false
-                                                }
-                                                checkBoxVisible.value = false
-                                            }
-
-                                            Icon(
-                                                painter = painterResource(id = R.drawable.baseline_share_24)
-                                            ) {
-                                                val selectedImageList = mediaList.filter { it.selected.value }
-
-                                                if (selectedImageList.isNotEmpty()) {
-
-                                                    val uriList = ArrayList<Uri>()
-
-                                                    uriList.addAll(selectedImageList.map { it.uri })
-
-                                                    shareImage(uriList)
-                                                }
-                                            }
-
-                                            Icon(
-                                                imageVector = Icons.Rounded.Delete
-                                            ) {
-                                                val selectedImageList = mediaList.filter { it.selected.value }
-
-                                                if (selectedImageList.isNotEmpty()) {
-
-                                                    deleteInProgress = true
-                                                    counterImageToDelete = selectedImageList.size
-
-                                                    cor {
-                                                        selectedImageList.forEach {
-                                                            setVarsAndDeleteImage(it.uri)
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
+                                            else Color.Transparent
+                                    )
                                 }
-                            }
+                            )
                         }
                     }
                 }
@@ -386,16 +271,6 @@ class MainActivity : ComponentActivity() {
         deleteAction = actionAfterDelete
 
         contentResolver.deletePhotoFromExternalStorage(photoUri, intentSenderLauncher)
-    }
-
-    private fun shareImage(imageUriArray: ArrayList<Uri>) {
-
-        val intent = Intent(Intent.ACTION_SEND_MULTIPLE)
-
-        intent.type = "image/*"
-        intent.putExtra(Intent.EXTRA_STREAM, imageUriArray)
-
-        startActivity(Intent.createChooser(intent, "Share Via"))
     }
 
     override fun onDestroy() {
